Unblock CI (#219)
* Bump CI os and Python versions

The integration tests use Python 3.8 / ubuntu 20.04. The problem is
that k8s-snap tests use type specifiers that require Python 3.10.

For this reason, we'll switch to Python 3.10 / ubuntu 22.04.

* Install lxd using k8s-snap composite action

This simplifies the gh workflows and also applies the Docker
iptables workaround.

* Use self-hosted runners

k8s-snap switched to self-hosted CI runners as the default Github
ones didn't have enough resources.

We'll update the k8s-dqlite CI job to use self-hosted runners as
well when running k8s-snap integration tests.

* Fix r install

* Remove "get" checks

The e2e tests are currently failing because of this k8s-dqlite error:

  k8s.k8s-dqlite[1947]: time="2025-01-09T15:00:22Z" level=error
  msg="error while range on /registry/clusterroles/system:aggregate-to-view :
  unexpected limit: want 0, got 1"

A separate PR is removing this check: https://github.com/canonical/k8s-dqlite/pull/212

That being considered, we'll apply this change here to unblock
the CI.

* fix self-hosted runner tag