Collect core dumps (#229)
* Include debug symbols for golang and dqlite builds

We need the debug symbols to analyze core dumps or live processes
using debuggers such as gdb or delve.

The size increase is reasonable, so we'll include them by default.

* Get detailed inspection reports

* Collect core dumps from perf tests

We're experiencing some dqlite crashes. The collected core dumps
should allow us to pinpoint the issue.

* perf test: upload the snaps in case of failure

We'll upload the snaps in case of perf test failures so that we
can further analyse the snap or potential core dumps.