Continue the kube-burner loop even in case of failures
We've been hitting Dqlite errors such as "database is locked".
At the moment it's unclear if this is a transient error or if
the database is compromised.

To answer this, we'll make a small change to the perf test,
resuming the kube-burner loop even in case of failures and
propagating the exceptions after completing the loop.