Add "--profiling-dir" setting
k8s-dqlite already has a setting that exposes a go pprof http
endpoint. The problem is that net/http/pprof requires a
predetermined duration, specifying for how long to collect
profiling data.

We can't know for sure how long the perf tests are going to last
and we'd like to collect profiling data for the entire test
duration.

For this reason, we'll add another setting called "--profiling-dir".
If set, we'll collect cpu time pprof data for the entire duration
of the k8s-dqlite process.

In the future, we may collect other pprof data (e.g. memory
statistics, etc).