Use list ranges instead of prefixes
k8s-dqlite can now properly use ranges, so we no longer need to use
fragile prefix logic.

Note that in order to properly support etcd range semantics, we
need to change the type of the "key" and "limit" arguments to []byte.
This means that we'll need to perform a cast as part of the SQL
query.

For now we'll only change the List and Count methods. For consistency,
we could update the other methods as well in a follow-up patch.